requirements.txt
exifread<3.0.0
pillow
pillow_heif
